Codex Routing
We look up the codex reference data.
If its a space poi then we set a space payload and endpoint
If it is a surface POI then we use the surface payload and endpoint
Otherwise we shove it in the codexreport
