+++ canonn/codex.py
@@ -2,84 +2,113 @@ import threading
 import requests
 import sys
 import json
+from emitter import Emitter
 
-class Codex(threading.Thread):
-    '''
-        Should probably make this a heritable class as this is a repeating pattern
-    '''
-
-    excludecodices={}
+class codexEmitter(Emitter):
+    types={}
     reporttypes={}
-
+    excludecodices={}
+    
     def __init__(self,cmdr, is_beta, system, x,y,z, entry, body,lat,lon,client):
-        threading.Thread.__init__(self)
-        self.cmdr = cmdr
-        self.system = system
+        Emitter.__init__(self,cmdr, is_beta, system, x,y,z, entry, body,lat,lon,client)
+        self.modelreport="xxreports"
+        self.modeltype="xxtypes"
         
-        self.x=x
-        self.y=y
-        self.z=z
-        self.body = body
-        self.lat = lat
-        self.lon = lon
-        self.is_beta = is_beta
-        self.entry = entry.copy()
-        self.client = client
+    def getSystemPayload(self,name):
+        payload=self.setPayload()
+        payload["userType"]="pc"
+        payload["reportType"]="new"
+        payload["type"]=name
+        payload["reportStatus"]="pending"
+        payload["isBeta"]=self.is_beta
+        payload["clientVersion"]=self.client
+    
+        return payload           
 
-    def run(self):
+    def getBodyPayload(self,name):
+        payload=self.getSystemPayload(name)
+        payload["bodyName"]=self.body
+        payload["coordX"]=self.x
+        payload["coordY"]=self.y
+        payload["coordZ"]=self.z
+        payload["latitude"]=self.lat
+        payload["longitude"]=self.lon
+    
+        return payload           
+    
+    def getCodexPayload(self):
+        payload=self.getBodyPayload(self.entry.get("Name"))
+        payload["entryId"]=self.entry.get("EntryID")
+        payload["codexName"]=self.entry.get("Name")
+        payload["codexNameLocalised"]=self.entry.get("Name_Localised")
+        payload["subCategory"]=self.entry.get("SubCategory")
+        payload["subCategoryLocalised"]=self.entry.get("SubCategory_Localised")
+        payload["category"]=self.entry.get("Category")
+        payload["categoryLocalised"]=self.entry.get("Category_Localised")
+        payload["regionName"]=self.entry.get("Region")
+        payload["regionLocalised"]=self.entry.get("Region_Localised")
+        payload["systemAddress"]=self.entry.get("SystemAddress")
+        payload["voucherAmount"]=self.entry.get("VoucherAmount")
+        payload["rawJson"]=self.entry
+        del payload["type"]
+        del payload["reportStatus"]
+        del payload["userType"]
+        del payload["reportType"]
         
-        if not Codex.excludecodices:
+        return payload
+        
+    
+    def getReportTypes(self,id):
+        if not codexEmitter.reporttypes.get(id):        
+            url="https://api.canonn.tech:2053/reporttypes?journalID={}".format(id)
+            print(url)
+            r=requests.get("https://api.canonn.tech:2053/reporttypes?journalID={}".format(id))    
+            if r.status_code == requests.codes.ok:
+
+                for exc in r.json():
+
+                    codexEmitter.reporttypes["{}".format(exc["journalID"])]={ "endpoint": exc["endpoint"], "location": exc["location"], "type": exc["type"]}
+                    print("get reporttype")
+                    print(codexEmitter.reporttypes.get("1400306")) 
+            else:
+                print("error")
+                
+                    
+    def getExcluded(self):
+        if not codexEmitter.excludecodices:
             r=requests.get("https://api.canonn.tech:2053/excludecodices")  
             if r.status_code == requests.codes.ok:
                 for exc in r.json():
-                    Codex.excludecodices["${}_name;".format(exc["codexName"])]=True
-                #print Codex.excludecodices
-
-        if not Codex.reporttypes:        
-            r=requests.get("https://api.canonn.tech:2053/reporttypes")    
-            if r.status_code == requests.codes.ok:
-                for exc in r.json():
-                    Codex.reporttypes["".format(exc["journalID"])]={ "endpoint": exc["endpoint"], "location": exc["location"]}
-                    print(Codex.reporttypes)
-                
-        if not Codex.excludecodices.get(self.entry.get("codexName")):
-            payload={}
-            payload["cmdrName"]=self.cmdr  
-            payload["systemName"]=self.system
-            payload["bodyName"]=self.body
-            payload["coordX"]=self.x
-            payload["coordY"]=self.y
-            payload["coordZ"]=self.z
-            payload["latitude"]=self.lat
-            payload["longitude"]=self.lon
-            payload["entryId"]=self.entry.get("EntryID")
-            payload["codexName"]=self.entry.get("Name")
-            payload["codexNameLocalised"]=self.entry.get("Name_Localised")
-            payload["subCategory"]=self.entry.get("SubCategory")
-            payload["subCategoryLocalised"]=self.entry.get("SubCategory_Localised")
-            payload["category"]=self.entry.get("Category")
-            payload["categoryLocalised"]=self.entry.get("Category_Localised")
-            payload["regionName"]=self.entry.get("Region")
-            payload["regionLocalised"]=self.entry.get("Region_Localised")
-            payload["systemAddress"]=self.entry.get("SystemAddress")
-            payload["voucherAmount"]=self.entry.get("VoucherAmount")
-            payload["rawJson"]=self.entry
-            payload["isBeta"]=self.is_beta
-            payload["clientVersion"]=self.client
-
-            try:        
-                r=requests.post("https://api.canonn.tech:2053/codexreports",data=json.dumps(payload),headers={"content-type":"application/json"})  
-            except:
-                print("[EDMC-Canonn] Issue posting codex entry " + str(sys.exc_info()[0]))                            
-                print r
+                    codexEmitter.excludecodices["${}_name;".format(exc["codexName"])]=True
+                    
+    def run(self):
+    
         
-def matches(d, field, value):
-	return field in d and value == d[field]	        
+        self.getExcluded()
+        
+        #is this a code entry and do we want to record it?
+        if self.entry["event"] == "CodexEntry" and not codexEmitter.excludecodices.get(self.entry.get("codexName")):
+            self.getReportTypes(self.entry.get("EntryID"))    
+            url=self.getUrl()
             
-'''
-    from canonn import journaldata
-    journaldata.submit(cmdr, system, station, entry)
-'''
+            jid=self.entry.get("EntryID")
+            reportType = codexEmitter.reporttypes.get(str(jid))
+            
+            if reportType:
+                print(reportType)
+                if reportType.get("location") == "body":
+                    payload=self.getBodyPayload(reportType.get("type"))
+                    self.modelreport=reportType.get("endpoint")
+                else:
+                    payload=self.getSystemPayload(reportType.get("type"))
+                    self.modelreport=reportType.get("endpoint")
+            else:
+                payload=self.getCodexPayload()
+                self.modelreport="codexreports"
+                   
+            print ("Send Reports {}/{}".format(url,self.modelreport))
+            
+            self.send(payload,url)
+            
 def submit(cmdr, is_beta, system, x,y,z, entry, body,lat,lon,client):
-    if entry["event"] == "CodexEntry":
-        Codex(cmdr, is_beta, system, x,y,z,entry, body,lat,lon,client).start()   
+    codexEmitter(cmdr, is_beta, system, x,y,z,entry, body,lat,lon,client).start()
