+++ load.py
@@ -8,10 +8,7 @@ from canonn import factionkill
 from canonn import nhss
 from canonn import codex
 from canonn import hdreport
-from canonn import btreports
-from canonn import fgreports
-from canonn import csreports
-from canonn import emitter
+
 
 
 import ttk
@@ -106,9 +103,7 @@ def journal_entry_wrapper(cmdr, is_beta, system, station, entry, state,x,y,z,bod
     #btreports.submit(cmdr, is_beta, system, x,y,z, entry, body,lat,lon,client)
     #fgreports.submit(cmdr, is_beta, system, x,y,z, entry, body,lat,lon,client)
     #csreports.submit(cmdr, is_beta, system, x,y,z, entry, body,lat,lon,client)
-    emitter.submit(cmdr, is_beta, system, x,y,z, entry, body,lat,lon,client)
-    
-    
+    codex.submit(cmdr, is_beta, system, x,y,z, entry, body,lat,lon,client)
     journaldata.submit(cmdr, is_beta, system, station, entry,client)
     
 def dashboard_entry(cmdr, is_beta, entry):
